requirements.txt
streamlit>=1.28.0
google-generativeai==0.8.3
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3 